pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "duckhunt-win"
version = "0.10.0"
description = "Prevent RubberDucky and keystroke injection attacks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Pedro M. Sosa" },
]
maintainers = [
    { name = "qb20nh" },
]
keywords = ["security", "rubberducky", "keystroke-injection", "usb-security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "pynput==1.8.1",
    "pystray==0.19.5",
    "Pillow==12.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller==6.17.0",
    "build==1.3.0",
]

[project.scripts]
duckhunt-win = "duckhunt_win.__main__:main"

[project.gui-scripts]
duckhunt-gui = "duckhunt_win.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["duckhunt_win"]

[tool.hatch.build.targets.sdist]
include = [
    "duckhunt_win/resources/*",
]

[tool.hatch.build]
include = [
    "duckhunt_win/resources/*",
]
